requirements.txt
discord.py
yfinance
python-dotenv